- Fixed Crons export bug - Higher Sec. Checking the users before they can accept a chat delete

+++ src/components/message.tsx
@@ -144,7 +144,11 @@ export const Message = ({
     };
 
   const deleteAllMessagesinChat = (chatId: string) => async () => {
-    await deleteAllMessagesInChat({ chatId });
+    if (message.type == "request") {
+      await deleteAllMessagesInChat({ chatId });
+    } else {
+      console.error("Error");
+    }
   };
 
   useEffect(() => {
